f414e150 Merge remote-tracking branch 'pvvtk/pv-master' into HEAD
cff717e6 Merge topic 'pvweb-add-scalar-range-controls' into master
14a81501 Merge topic 'one-step-ios-build' into master
a89b405d Merge topic 'disable_opengl_error_checks_in_release_builds' into master
33b149bd Merge topic 'Progress-related-events' into master
89c6cd22 Merge topic 'stringcolors' into master
b6e2cf34 Add valid image for OpenGL2
3b53ef0d Merge topic 'fix-vtkMakeInstantiator-regression' into master
751982cf Merge topic 'fix_texture_failures' into master
6b86fb05 Fix inverted ifdef NDEBUG in commit f1a90fd2.
85b09ec6 Merge topic 'disable_opengl_error_checks_in_release_builds' into master
ca6c800d Fixed when mappers try to reuse the same texture
84113485 Merge topic 'compile_error' into master
b6e796d7 Fix dashbaord compile error and some long lines
9ea80200 In the vtkweb loader, add css specific to the proxy editor widget.
98e3bac5 Merge topic 'pvweb-document-disable-registry' into master
...

+++ Common/Misc/Testing/Cxx/UnitTestFunctionParser.cxx
@@ -270,7 +270,7 @@ int TestVectors()
       {
       if (!vtkMathUtilities::FuzzyCompare(
             a[j], b[j],
-            std::numeric_limits<double>::epsilon()*(double)1.0))
+            std::numeric_limits<double>::epsilon() * 1.0))
         {
         std::cout << " Cross expected " << a[j]
                   << " but got " << b[j];
@@ -285,7 +285,7 @@ int TestVectors()
     double dot = parser->GetScalarResult();
     if (!vtkMathUtilities::FuzzyCompare(
           dot, 0.0,
-          std::numeric_limits<double>::epsilon()*(double)1.0))
+          std::numeric_limits<double>::epsilon() * 1.0))
       {
       std::cout << " Dot " << 0.0
                 << " but got " << dot;
@@ -300,7 +300,7 @@ int TestVectors()
     double mag = parser->GetScalarResult();
     if (!vtkMathUtilities::FuzzyCompare(
           mag, 1.0,
-          std::numeric_limits<double>::epsilon()*(double)2.0))
+          std::numeric_limits<double>::epsilon() * 2.0))
       {
       std::cout << " Mag expected" << 1.0
                 << " but got " << mag;
